Convert graphile-utils to typescript
