+++ packages/graphile-utils/__tests__/ExtendSchemaPlugin.test.js
@@ -3,7 +3,7 @@ import {
   makeAddInflectorsPlugin,
   gql,
   embed,
-} from "../src";
+} from "../";
 import {
   buildSchema,
   // defaultPlugins,
