--- packages/graphile-utils/src/makeAddInflectorsPlugin.js
@@ -1,7 +1,0 @@
-export default function makeAddInflectorsPlugin(additionalInflectors) {
-  return builder => {
-    builder.hook("inflection", (inflection, build) => {
-      return build.extend(inflection, additionalInflectors);
-    });
-  };
-}
