+++ packages/graphile-utils/src/makeAddInflectorsPlugin.ts
@@ -1,0 +1,9 @@
+import { Plugin } from 'graphile-build';
+
+export default function makeAddInflectorsPlugin(additionalInflectors: {[str: string]: ((...args: any[]) => any)}): Plugin {
+  return builder => {
+    builder.hook("inflection", (inflection, build) => {
+      return build.extend(inflection, additionalInflectors);
+    });
+  };
+}
